+++ Modules/Ruleset.Logging/Public/Write-LogFile.ps1
@@ -79,7 +79,7 @@ PS> [hashtable] $HashResult = Get-SomeHashTable
 PS> Write-LogFile -Path "C:\logs" -LogName "Settings" -Tags "MyTag" -Hash $HashResult
 PS> $HeaderStack.Pop() | Out-Null
 
-Will write entry $HashResult to log C:\logs\Settings_15.12.20.log with a header set to "My Header"
+Will write entire $HashResult to log C:\logs\Settings_15.12.20.log with a header set to "My Header"
 
 .EXAMPLE
 PS> $HeaderStack.Push("My Header")