+++ Modules/Ruleset.Utility/Public/Set-NetworkProfile.ps1
@@ -36,6 +36,10 @@ Set network profile for physical network interfaces
 Set network profile for each physical/hardware network interfaces
 Recommended is "Public" profile for maximum security, unless "Private" is needed
 
+.PARAMETER NetworkCategory
+Specify network category which to apply to all NIC's.
+If not specified, you're prompted for each NIC individually
+
 .EXAMPLE
 PS> Set-NetworkProfile
 
@@ -46,7 +50,7 @@ None. You cannot pipe objects to Set-NetworkProfile
 None. Set-NetworkProfile does not generate any output
 
 .NOTES
-TODO: It looks like option to change network profile in settings app will be gone after using this
+HACK: It looks like option to change network profile in settings app will be gone after using this
 function in an elevated prompt
 #>
 function Set-NetworkProfile
@@ -54,12 +58,17 @@ function Set-NetworkProfile
 	[CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = "High",
 		HelpURI = "https://github.com/metablaster/WindowsFirewallRuleset/blob/master/Modules/Ruleset.Utility/Help/en-US/Set-NetworkProfile.md")]
 	[OutputType([void])]
-	param ()
+	param (
+		[Parameter()]
+		[ValidateSet("Public", "Domain", "Private")]
+		[string] $NetworkCategory
+	)
 
 	Write-Debug -Message "[$($MyInvocation.InvocationName)] ParameterSet = $($PSCmdlet.ParameterSetName):$($PSBoundParameters | Out-String)"
 
 	if ($PSCmdlet.ShouldProcess("Configure network profile"))
 	{
+		# NOTE: This will include external switches bound to physical adapters as well
 		[string[]] $HardwareInterfaces = Get-NetConnectionProfile |
 		Select-Object -ExpandProperty InterfaceAlias
 
@@ -67,54 +76,58 @@ function Set-NetworkProfile
 		# TODO: When could second check be true? (interfaces -eq 0)
 		if (!$HardwareInterfaces -or ($HardwareInterfaces.Length -eq 0))
 		{
-			# TODO: we should base this on IPv*Connectivity given by Get-NetConnectionProfile
+			# TODO: We should base this on IPv*Connectivity given by Get-NetConnectionProfile
 			Write-Warning -Message "Unable to set network profile, machine not connected to network"
 			return
 		}
-
-		# NOTE: not logging this warning
-		Write-Warning -Message "For maximum security choose 'Public' network profile"
 
 		foreach ($Interface in $HardwareInterfaces)
 		{
 			Write-Debug -Message "[$($MyInvocation.InvocationName)] Processing $Interface"
 
-			# User prompt default values
-			[int32] $Default = 0
-			[ChoiceDescription[]] $Choices = @()
-			$Public = [ChoiceDescription]::new("&Public")
-			$Private = [ChoiceDescription]::new("P&rivate")
-			$Abort = [ChoiceDescription]::new("&Abort")
-
-			$Public.HelpMessage = "Your PC is hidden from other devices on the network"
-			$Private.HelpMessage = "Your PC is discoverable and can be used for file and printer sharing"
-			$Abort.HelpMessage = "Abort operation, no change is done to network profile"
-
-			$Choices += $Public
-			$Choices += $Private
-			$Choices += $Abort
-
-			$Title = "Choose network profile"
-			$Question = "Which network profile to set for '$Interface' interface?"
-			$Decision = $Host.UI.PromptForChoice($Title, $Question, $Choices, $Default)
-
-			if ($Decision -eq 0)
+			if ($NetworkCategory)
 			{
-				$NetworkCategory = "Public"
-			}
-			elseif ($Decision -eq 1)
-			{
-				$NetworkCategory = "Private"
+				$Category = $NetworkCategory
 			}
 			else
 			{
-				Write-Warning -Message "The operation has been canceled by the user"
-				return
+				# User prompt default values
+				[int32] $Default = 0
+				[ChoiceDescription[]] $Choices = @()
+				$Public = [ChoiceDescription]::new("&Public")
+				$Private = [ChoiceDescription]::new("P&rivate")
+				$Abort = [ChoiceDescription]::new("&Abort")
+
+				$Public.HelpMessage = "Your PC is hidden from other devices on the network"
+				$Private.HelpMessage = "Your PC is discoverable and can be used for file and printer sharing"
+				$Abort.HelpMessage = "Abort operation, no change is done to network profile"
+
+				$Choices += $Public
+				$Choices += $Private
+				$Choices += $Abort
+
+				$Title = "Choose network profile"
+				$Question = "Which network profile to set for '$Interface' interface?"
+				$Decision = $Host.UI.PromptForChoice($Title, $Question, $Choices, $Default)
+
+				if ($Decision -eq 0)
+				{
+					$Category = "Public"
+				}
+				elseif ($Decision -eq 1)
+				{
+					$Category = "Private"
+				}
+				else
+				{
+					Write-Warning -Message "The operation has been canceled by the user"
+					return
+				}
 			}
 
-			Set-NetConnectionProfile -InterfaceAlias $Interface -NetworkCategory $NetworkCategory
+			Set-NetConnectionProfile -InterfaceAlias $Interface -NetworkCategory $Category
 			Write-Information -Tags $MyInvocation.InvocationName `
-				-MessageData "INFO: Network profile set to '$NetworkCategory' for '$Interface' interface"
+				-MessageData "INFO: Network profile set to '$Category' for '$Interface' interface"
 		}
 	}
 	else