+++ Test/MarkdownLinkTest.ps1
@@ -60,7 +60,7 @@ param (
 . $PSScriptRoot\..\Config\ProjectSettings.ps1 $PSCmdlet
 
 if (Approve-Execute -Title "Markdown links" -Question "Run markdown link test?" -Force:$Force `
-		-Accept "Run markdown link test on entry repository" -Deny "Skip link test operation")
+		-Accept "Run markdown link test on entire repository" -Deny "Skip link test operation")
 {
 	Test-MarkdownLinks $ProjectRoot -Recurse -Log
 }