+++ Readme/CHANGELOG.md
@@ -75,7 +75,7 @@ If you would like to see fresh changes done since last release you can do so on 
   - `Test-TargetComputer` renamed to `Test-Computer` now experimental and no longer useful
   - `Resolve-Host` function improved, removed deprecated API, duplicate code etc.
   - Registry drilling functions updated for remote registry
-  - `ConvertFrom-OSBuild` updated for 21H1 and 21H2 OS version (including Windows 11)
+  - `ConvertFrom-OSBuild` updated for 21H1 and 21H2 OS version (including Windows 11 and server 2022)
 
 - Documentation